requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-WTF==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
passlib==1.7.4
pycparser==2.21
pymongo==4.0.1
python-dotenv==0.19.2
Werkzeug==2.0.2
WTForms==3.0.1
